Change permission to blessing. The rights situation is ambiguous.
Michael is happy with this project, and has given permission as much as
he is able; sadly, the actual rights to the book are ambiguous, with
the publisher going out of business and the PDF being released by Dr.
Dobbs Journal.

Removed the word permission as to not imply that Michael Abrash has the
ability to grant legal permission. He doesn't.